pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adr-agent"
version = "0.1.0"
description = "Per-repository architectural memory for AI agents"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "python-frontmatter>=1.1",
    "anthropic>=0.40",
    "pyyaml>=6.0",
]

[project.scripts]
adr-agent = "adr_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/adr_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.12",
]
